requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
streamlit>=1.28.0
pillow>=10.0.0

